refactor: move LinkPreviewResolutionService to Application/Services (#356)
It's a cross-cutting service, not a feature slice. Used by both
Channel and Conversation SendMessage handlers.
